Fix migrations functions
