+++ apps/backend-ops/src/firestoreMigrations/index.ts
@@ -1,6 +1,5 @@
 import * as v0001 from './0001';
 import * as v0002 from './0002';
-import * as v0004 from './0004';
 import { Firestore } from '../admin';
 
 export interface IMigration {
@@ -13,8 +12,7 @@ export interface IMigrationWithVersion extends IMigration {
 
 export const MIGRATIONS = {
   1: v0001,
-  2: v0002,
-  4: v0004,
+  2: v0002
 };
 
 export const VERSIONS_NUMBERS = Object.keys(MIGRATIONS).map(s => parseInt(s, 10));
